requirements.txt
Chameleon==2.9.2
Mako==0.7.1
MarkupSafe==0.15
PasteDeploy==1.5.0
PasteScript==1.7.5
WebOb==1.2
pyramid==1.3.2
pyramid_debugtoolbar==1.0.2
-e git://github.com/breckprog/pyramid_paas.git#egg=pyramid_paas
repoze.lru==0.5
translationstring==1.1
venusian==1.0a6
waitress==0.8.1
wsgiref==0.1.2
zope.deprecation==4.0.0
zope.interface==4.0.1
sqlalchemy==0.7.8
psycopg2==2.4.5
